Add custom tags and allow the script to filter items based on tags
The new tags for online-starter and online-professional are added to selected
templates and imagestreams. The script is modified to accept the custom tags via
flag '-t' or '--tags' and custom tags are concatenated but separated by comma.

If custom tags are provided, all templates and imagestreams will be imported under
the folder with the name of the tag(s) and under /templates or /imagestreams
respectively. There are not further sub-directori(es).

The index.json and README.rd are also added into the folder(s) to provide further
information in case users need to locate the file(s) or source(s).

The script is now able to detect file collision and will append index numer to
file name to avoid file overwritten issue.

Updated the README to reflect new Online tags for YAML files and further details.

Signed-off-by: Vu Dinh <vdinh@redhat.com>