Update ImageStreamTag name to version only
This is a copy of https://github.com/jboss-container-images/openjdk/pull/103 which was a problem for OCP 4.3 (see https://bugzilla.redhat.com/show_bug.cgi?id=1778613)

I've targetted `master` but I'm not sure which versions of OCP are affected nor which branches are relevant to those. Can you recommend whether we need to merge to other places too?